% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ocgrtools-package.r
\docType{package}
\name{Ocgrtools}
\alias{Ocgrtools}
\alias{Ocgrtools-package}
\title{Ocgrtools: Visualize oceanographic datasets}
\description{
Readlily plot depth profiles of ocean properties and utilize a dataset of various
properties collected as a part of an expedition in the subarctic NE Pacific
(LineP cruise) that is contained in the package
}

