% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplot.R
\name{pplot}
\alias{pplot}
\title{pplot}
\usage{
pplot(df, property, depth)
}
\arguments{
\item{df}{dataframe}

\item{property}{numeric vector}

\item{depth}{numeric vector}
}
\value{
plot
}
\description{
plot ocean property with depth
}
\examples{
#' pplot(data,"temp","depth")
}

